apps/api/pyproject.toml
[project]
name = "timestamp-msi-api"
version = "0.1.0"
description = "FastAPI backend for the Timestamp_msi TCGA CRC MSI workstation."
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "pandas>=2.0.0",
  "python-multipart>=0.0.9",
  "pydantic-settings>=2.5.0",
  "sqlalchemy>=2.0.0",
  "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["storage*", "tests*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 88
target-version = "py310"
